# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SimpleShop
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  version: 1.0.0
...
servers:
  - url: http://localhost:5005
...
paths:
...
  /products:
    get:
      tags:
        - product
      summary: Get All Products
      operationId: getProducts
      responses:
        '200':
          description: successful receved products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    post:
      tags:
        - product
      summary: Create Product
      description: This can only be done by the logged in admin.
      operationId: createProduct
      requestBody:
        description: Created product object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductDummy'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Product dose already exist.
  /products/{id}:
    get:
      tags:
        - product
      summary: Get product by ID
      description: This can only be done by the logged in admin.
      operationId: getProduct
      parameters:
        - name: id
          in: path
          description: public_id of Product that need to be updated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product dose not exist.
        '405':
          description: Function not allowed
    put:
      tags:
        - product
      summary: edit product
      description: This can only be done by the logged in admin.
      operationId: editProduct
      parameters:
        - name: id
          in: path
          description: public_id of Product that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Created product object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductDummy'
      responses:
        '200':
          description: successful deleted Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product dose not exist.
        '405':
          description: Function not allowed
    delete:
      tags:
        - product
      summary: delete product
      description: This can only be done by the logged in admin.
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          description: public_id of Product that need to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product dose not exist.
        '405':
          description: Function not allowed
  # SUPPLIER
  /supplier:
    get:
      tags:
        - supplier
      summary: Get All Suppliers
      operationId: getSuppliers
      responses:
        '200':
          description: successful receved Suppliers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
    post:
      tags:
        - supplier
      summary: Create Supplier
      description: This can only be done by the logged in admin.
      operationId: createSupplier
      requestBody:
        description: Created Supplier object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SupplierDummy'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '400':
          description: Supplier dose already exist.
  /supplier/{id}:
    get:
      tags:
        - supplier
      summary: Get Supplier by ID
      description: This can only be done by the logged in admin.
      operationId: getSupplier
      parameters:
        - name: id
          in: path
          description: public_id of Supplier that need to be updated
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '404':
          description: Supplier dose not exist.
        '405':
          description: Function not allowed
    put:
      tags:
        - supplier
      summary: edit Supplier
      description: This can only be done by the logged in admin.
      operationId: editSupplier
      parameters:
        - name: id
          in: path
          description: public_id of Supplier that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Created Supplier object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SupplierDummy'
      responses:
        '200':
          description: successful deleted Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '404':
          description: Supplier dose not exist.
        '405':
          description: Function not allowed
    delete:
      tags:
        - supplier
      summary: delete Supplier
      description: This can only be done by the logged in admin.
      operationId: deleteSupplier
      parameters:
        - name: id
          in: path
          description: public_id of Supplier that need to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful deleted Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
        '404':
          description: Supplier dose not exist.
        '405':
          description: Function not allowed
  /supplier/{id}/product/add:
    put:
      tags:
        - supplier
      summary: add Product to Supplier
      description: This can only be done by the logged in admin.
      operationId: addProductToSupplier
      parameters:
        - name: id
          in: path
          description: public_id of Supplier that need to be updated
          required: true
          schema:
            type: string
            example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
      requestBody:
        description: Created Supplier object
        content:
          multipart/form-data:
            schema:
              properties:
                product_id:
                  type: string
                  example: "522369cd-2185-442e-a88e-9dwetq9c2386"
      responses:
        '200':
          description: successful deleted Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierProduct'
        '404':
          description: Supplier dose not exist.
        '405':
          description: Function not allowed
  /supplier/{id}/product/{s_p_id}:
    put:
      tags:
        - supplier
      summary: delete Product to Supplier
      description: This can only be done by the logged in admin.
      operationId: addProductToSupplier
      parameters:
        - name: id
          in: path
          description: public_id of Supplier that need to be updated
          required: true
          schema:
            type: string
            example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
        - name: s_p_id
          in: path
          description: public_id of Supplier-Product that need to be deleted
          required: true
          schema:
            type: string
            example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
      requestBody:
        description: Created Supplier object
        content:
          multipart/form-data:
            schema:
              properties:
                product_id:
                  type: string
                  example: "522369cd-2185-442e-a88e-9dwetq9c2386"
      responses:
        '200':
          description: successful deleted Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierProduct'
        '404':
          description: Supplier dose not exist.
        '405':
          description: Function not allowed

  # ORDER
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        public_id:
          type: string
          example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
        title:
          type: string
          example: Super Supplier
    SupplierDummy:
      type: object
      properties:
        title:
          type: string
          example: Super Product
    SupplierProduct:
      type: object
      properties:
        supplier_id:
          type: string
          example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
        product_id:
          type: string
          example: '522369cd-2185-442e-a88e-9dwetq9c2386'
    # PRODUCT
    Product:
      type: object
      properties:
        public_id:
          type: string
          example: '525079cd-2185-4e2e-a88e-9d2b5e9c2b81'
        title:
          type: string
          example: Super Product
        price:
          type: number
          example: 1.3
        stock:
          type: integer
          example: 30
    ProductDummy:
      type: object
      properties:
        title:
          type: string
          example: Super Product
        price:
          type: number
          example: 1.3
        stock:
          type: integer
          example: 30
    # USER
...